modified files for PA3

+++ PA3/include/knngpu.h
@@ -8,6 +8,6 @@
 
 
 // This is the declaration of the function that will execute on the GPU.
-__global__ void knngpu(int *, int *, int);
+__global__ void knngpu(float *, float *, int);
 
 #endif // KNNGPU_H_
